Move H5DenseMatrixWriter to cpp file
